progress: simplify types and blocks
