+++ src/modules/AieraChat/Messages/MessageFactory/Block/index.tsx
@@ -1,12 +1,42 @@
 import React from 'react';
 import { match } from 'ts-pattern';
-import { List } from './List';
-import { Quote } from './Quote';
-import { Table } from './Table';
-import { Text } from './Text';
-import { BlockType, ContentBlock } from './types';
-import { Image } from './Image';
-import { Chart } from './Chart';
+import { ChartBlock, Chart } from './Chart';
+import { ImageBlock, Image } from './Image';
+import { ListBlock, List } from './List';
+import { QuoteBlock, Quote } from './Quote';
+import { TableBlock, Table } from './Table';
+import { TextBlock, Text } from './Text';
+
+// Citation type for referencing sources
+type Citation = {
+    id: string;
+    text: string;
+    source: string;
+    url?: string;
+    date?: string;
+    author?: string;
+};
+
+// Type for content that can include citations
+export type CitableContent = (string | Citation)[];
+
+export enum BlockType {
+    text = 'text',
+    table = 'table',
+    list = 'list',
+    image = 'image',
+    quote = 'quote',
+    chart = 'chart',
+}
+
+// Base type for all content blocks
+export interface BaseBlock {
+    id: string;
+    type: BlockType;
+}
+
+// Union type for all content blocks
+export type ContentBlock = TextBlock | TableBlock | ListBlock | ImageBlock | QuoteBlock | ChartBlock;
 
 export function Block(props: ContentBlock) {
     return match(props)
